refactor(baseball): 한 클래스 내에 있던 기능을 나눔
